Create component for showing appointements
