+++ .prettierrc
@@ -1,0 +1,5 @@
+{
+    "trailingComma": "all",
+    "singleQuote": true,
+    "arrowParens": "avoid"
+  }
